feat(core): extract @ixo/oracle-core — the portable, web-standard harness core
The appraisal rejected a ./core subpath inside the Node package: installing
it still resolves the full Node/Nest/native dependency surface. This lands
the real package instead. @ixo/oracle-core carries the modules that are
already web-standard — the authority-kernel primitives (audit, permissions,
budgets, ledger, execution broker), the transport-neutral turn seam, the
semantic-routing layer, the operator model-policy layer (policy, credential
broker, adapter registry, builtin defaults), the typed event names +
envelope, and the emoji util — with dependencies limited to langchain,
@langchain/core, zod, node-emoji (pure JS), and @ixo/oracles-events
(eventemitter2 only, audited).

@ixo/oracle-runtime becomes the Node/Nest adapter: every old import path is
a re-export shim over the core package, so the public 1.x surface (and all
error-class identities) is unchanged — proven by the entire 953-test suite
passing through the shims without a single test edit. The canonical Logger
type and EVENT_NAMES move to core; plugin-api and the scoped emitter
re-export them. Node-coupled surfaces stay in the runtime deliberately:
context-guard (full RuntimeContext shape), llm-provider (provider SDKs),
scoped-emitter (ambient wiring).

Also in core, foundation-scoped config envelopes: signedConfigEnvelopeSchema
(content-addressed, controller-authorized, anti-rollback fields, hash-chain
link, runtime channel) and dataPolicySchema (classifications, jurisdictions,
processors, custody, retention, per-classification placements with
fail-closed isPlacementPermitted, logging-off defaults). Schemas + verifier
interfaces only — the Phase 5 loader consumes them.

CI gains a core portability gate: an esbuild neutral-platform bundle with
packages external and NO nodejs_compat, failing on any node:* specifier —
compat stubs proving nothing is exactly the failure mode this avoids.
Lockfile updated surgically with only the two importer entries (validated
with pnpm install --frozen-lockfile --lockfile-only); esbuild pinned to an
already-resolved version to avoid dragging new platform binaries in.

Deferred to the next tranche (tracked in docs, not dropped): moving graph/,
registries/, plugin-api/, manifest/ into core — blocked on making
plugin-api's Nest route types structural and auditing prompt-composer's
@ixo/common dependency.
